move element warnings to actual use site
